changed weight to double to allow user to enter for example 25,5kg. modified workout view pages workout view

+++ lib/models/workoutMovement.dart
@@ -2,7 +2,7 @@ class WorkoutMovement {
   String movement;
   int sets;
   List<int> reps;
-  List<int> weights;
+  List<double> weights;
 
   WorkoutMovement({
     required this.movement,
